Long term optimization (#4)
This patch implments the long-term CPU core scaling optimization that
1) to run a heuristics algorithm to decide whether RSS buckets need to be migrated and the number of normal cores;
2) to update the mapping from RSS buckets to normal CPU cores;
3) to call the NIC's function to update the RSS LB at the NIC hardware;

* Long term optimization
* Implement reta_table_to port
* Implement the cpu scaling algorithm
* Remove the port constraint

+++ core/modules/nfv_switch.h
@@ -134,7 +134,7 @@ class NFVSwitch final : public Module {
 
   // The number of normal / reserved CPU cores
   // |normal_core_count_| + |idle_core_count_| <= |total_core_count_|
-  int total_core_count_ = 0;
+  uint16_t total_core_count_ = 0;
   int normal_core_count_ = 0;
   int idle_core_count_ = 0;
